+++ pl_examples/domain_templates/gan.py
@@ -19,9 +19,8 @@ import torchvision.transforms as transforms
 from torch.utils.data import DataLoader
 from torchvision.datasets import MNIST
 
-from pytorch_lightning.core import LightningModule
-from pytorch_lightning.core import data_loader
-from pytorch_lightning.trainer import Trainer
+from pytorch_lightning.core import LightningModule, data_loader
+from pytorch_lightning.training import Trainer
 
 
 class Generator(nn.Module):
