+++ pytorch_lightning/training/callback_config.py
