+++ tests/trainer/test_trainer.py
@@ -23,7 +23,7 @@ from tests.models import (
     LightTestDataloader,
 )
 from pytorch_lightning.core.lightning import load_hparams_from_tags_csv
-from pytorch_lightning.trainer.logging import TrainerLoggingMixin
+from pytorch_lightning.training.logging import TrainerLoggingMixin
 from pytorch_lightning.utilities.debugging import MisconfigurationException
